move formatUpdate to updateView
